fix: preserve contour dtype in fit_quad to maintain numeric precision
The original code forced float32 conversion, which changed approxPolyDP's
numeric behavior on the 'already a clean quad' fast path, violating the
byte-for-byte identical behavior constraint. Dropping the explicit dtype
cast preserves int32 from cv2.findContours in classical-CV callers while
still accepting float arrays from YOLO.

Also adds test_detect_slots_is_deterministic_on_real_camera_folder to
catch coordinate-level drift in future refactors.

+++ src/slot_detection.py
@@ -103,7 +103,7 @@ def fit_quad(cnt, inset_px=6.0):
     to its minimum-area bounding rect. Either way, shrink the result inward
     so it doesn't overshoot the true painted line (see _shrink_polygon).
     """
-    cnt = np.asarray(cnt, dtype=np.float32).reshape(-1, 1, 2)
+    cnt = np.asarray(cnt).reshape(-1, 1, 2)
     peri = cv2.arcLength(cnt, True)
     approx = cv2.approxPolyDP(cnt, 0.02 * peri, True)
     if len(approx) == 4:
